langgraph/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
langgraph==0.6.7
langchain==0.3.27
pydantic==2.11.9
python-dotenv==1.1.1
requests==2.32.5
openai>=1.40.0
